+++ src/dev/testClass/client/TestExtension.ts
@@ -1,0 +1,3 @@
+import CoreKernelExtension from '../../../classes/CoreKernelExtension.js';
+
+export default class TestExtension extends CoreKernelExtension {}