+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './EntityValidator.js';
+export * from './Executable.js';
 export * from './XUtil.js';